BarsHelper and AccentColorHelper moved to namespace UWPHelper.UI
